fix Pixelcade install script / install_java

+++ packages/sx05re/emulators/freej2me/scripts/freej2me.sh
@@ -6,9 +6,11 @@
 . /etc/profile
 
 # Lets check if java is installed and up to date, called from profile
-install_java;
+install_java
 
-# We make sure we are using the provided jar
-cp -rf /usr/lib/libretro/freej2me-lr.jar ${HOME}/roms/bios
-
+if [[ -f "/storage/roms/bios/jdk/eeversion" ]]; then
+    cp -rf /usr/lib/libretro/freej2me-lr.jar ${HOME}/roms/bios
+else
+    exit 1
+fi
 exit 0
